+++ packages/core/src/mock/mockMatterbridge.js
@@ -4,7 +4,6 @@ import EventEmitter from 'node:events';
 export class MockMatterbridge extends EventEmitter {
   version = '1.0.0';
 
-  // eslint-disable-next-line @typescript-eslint/no-useless-constructor
   constructor() {
     super();
   }